PM / devfreq: Make the monitor workqueue high priority
Running the devfreq monitor in a normal-priority workqueue can result
in the devfreq mutex lock being held for a while when the monitor
worker sleeps inside update_devfreq(). Freeing up the devfreq mutex
lock is important to ensure timely devfreq boosts, so run the monitor
workqueue in a high-priority workqueue to reduce the lock duration.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I414c5d837da708bf406d3b8d8aa349c0a7c1760e
Signed-off-by: frost. (frostg-012) <frost@nusantararom.org>